+++ src/main/java/org/zalando/nakadi/controller/CursorsController.java
@@ -15,14 +15,14 @@ import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.context.request.NativeWebRequest;
 import org.zalando.nakadi.domain.ItemsWrapper;
 import org.zalando.nakadi.domain.NakadiCursor;
-import org.zalando.nakadi.exceptions.InternalNakadiException;
-import org.zalando.nakadi.exceptions.InvalidCursorException;
-import org.zalando.nakadi.exceptions.NakadiException;
-import org.zalando.nakadi.exceptions.NakadiRuntimeException;
-import org.zalando.nakadi.exceptions.NoSuchEventTypeException;
+import org.zalando.nakadi.exceptions.runtime.NakadiRuntimeException;
 import org.zalando.nakadi.exceptions.runtime.CursorsAreEmptyException;
 import org.zalando.nakadi.exceptions.runtime.FeatureNotAvailableException;
+import org.zalando.nakadi.exceptions.runtime.InternalNakadiException;
+import org.zalando.nakadi.exceptions.runtime.InvalidCursorException;
 import org.zalando.nakadi.exceptions.runtime.InvalidStreamIdException;
+import org.zalando.nakadi.exceptions.runtime.NoSuchEventTypeException;
+import org.zalando.nakadi.exceptions.runtime.NoSuchSubscriptionException;
 import org.zalando.nakadi.exceptions.runtime.RequestInProgressException;
 import org.zalando.nakadi.exceptions.runtime.ServiceTemporarilyUnavailableException;
 import org.zalando.nakadi.exceptions.runtime.UnableProcessException;
@@ -46,6 +46,8 @@ import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
+import static javax.ws.rs.core.Response.Status.INTERNAL_SERVER_ERROR;
+import static javax.ws.rs.core.Response.Status.NOT_FOUND;
 import static javax.ws.rs.core.Response.Status.SERVICE_UNAVAILABLE;
 import static org.springframework.http.ResponseEntity.noContent;
 import static org.springframework.http.ResponseEntity.ok;
@@ -81,7 +83,7 @@ public class CursorsController {
                     .map(cursor -> cursor.withToken(cursorTokenService.generateToken()))
                     .collect(Collectors.toList());
             return new ItemsWrapper<>(cursors);
-        } catch (final NakadiException e) {
+        } catch (final InternalNakadiException e) {
             throw new NakadiRuntimeException(e);
         }
     }
@@ -110,12 +112,15 @@ public class CursorsController {
             }
         } catch (final NoSuchEventTypeException | InvalidCursorException e) {
             return create(Problem.valueOf(UNPROCESSABLE_ENTITY, e.getMessage()), request);
+        } catch (final NoSuchSubscriptionException e) {
+            LOG.error("Subscription not found", e);
+            return create(Problem.valueOf(NOT_FOUND, e.getMessage()), request);
         } catch (final ServiceTemporarilyUnavailableException e) {
             LOG.error("Failed to commit cursors", e);
             return create(Problem.valueOf(SERVICE_UNAVAILABLE, e.getMessage()), request);
-        } catch (final NakadiException e) {
+        } catch (final InternalNakadiException e) {
             LOG.error("Failed to commit cursors", e);
-            return create(e.asProblem(), request);
+            return create(Problem.valueOf(INTERNAL_SERVER_ERROR, e.getMessage()), request);
         }
     }
 
@@ -131,8 +136,8 @@ public class CursorsController {
             throw new UnableProcessException(e.getMessage());
         } catch (final InvalidCursorException e) {
             return create(Problem.valueOf(UNPROCESSABLE_ENTITY, e.getMessage()), request);
-        } catch (final NakadiException e) {
-            return create(e.asProblem(), request);
+        } catch (final InternalNakadiException e) {
+            return create(Problem.valueOf(INTERNAL_SERVER_ERROR, e.getMessage()), request);
         }
     }