+++ src/main/java/org/zalando/nakadi/exceptions/runtime/CompactionException.java
@@ -1,0 +1,9 @@
+package org.zalando.nakadi.exceptions.runtime;
+
+public class CompactionException extends NakadiBaseException {
+
+    public CompactionException(final String message) {
+        super(message);
+    }
+
+}