+++ src/main/java/org/zalando/nakadi/exceptions/runtime/ConflictException.java
@@ -1,6 +1,6 @@
 package org.zalando.nakadi.exceptions.runtime;
 
-public class ConflictException extends MyNakadiRuntimeException1 {
+public class ConflictException extends NakadiBaseException {
 
     public ConflictException(final String message) {
         super(message);