+++ src/main/java/org/zalando/nakadi/exceptions/runtime/DuplicatedSubscriptionException.java
@@ -1,6 +1,6 @@
 package org.zalando.nakadi.exceptions.runtime;
 
-public class DuplicatedSubscriptionException extends MyNakadiRuntimeException1 {
+public class DuplicatedSubscriptionException extends NakadiBaseException {
 
     public DuplicatedSubscriptionException(final String msg, final Exception cause) {
         super(msg, cause);