+++ src/main/java/org/zalando/nakadi/exceptions/runtime/ErrorGettingCursorTimeLagException.java
@@ -1,9 +1,8 @@
-package org.zalando.nakadi.exceptions;
+package org.zalando.nakadi.exceptions.runtime;
 
 import org.zalando.nakadi.domain.NakadiCursor;
-import org.zalando.nakadi.exceptions.runtime.MyNakadiRuntimeException1;
 
-public class ErrorGettingCursorTimeLagException extends MyNakadiRuntimeException1 {
+public class ErrorGettingCursorTimeLagException extends NakadiBaseException {
 
     private final NakadiCursor failedCursor;