+++ src/main/java/org/zalando/nakadi/exceptions/runtime/EventTypeTimeoutException.java
@@ -1,6 +1,6 @@
 package org.zalando.nakadi.exceptions.runtime;
 
-public class EventTypeTimeoutException extends MyNakadiRuntimeException1 {
+public class EventTypeTimeoutException extends NakadiBaseException {
 
     public EventTypeTimeoutException(final String message) {
         super(message);