+++ src/main/java/org/zalando/nakadi/exceptions/runtime/InvalidEventTypeException.java
@@ -1,6 +1,6 @@
 package org.zalando.nakadi.exceptions.runtime;
 
-public class InvalidEventTypeException extends MyNakadiRuntimeException1 {
+public class InvalidEventTypeException extends NakadiBaseException {
 
     public InvalidEventTypeException(final String message) {
         super(message);