+++ src/main/java/org/zalando/nakadi/exceptions/runtime/InvalidLimitException.java
@@ -1,0 +1,8 @@
+package org.zalando.nakadi.exceptions.runtime;
+
+public class InvalidLimitException extends NakadiBaseException {
+
+    public InvalidLimitException(final String message) {
+        super(message);
+    }
+}