+++ src/main/java/org/zalando/nakadi/exceptions/runtime/NoStorageException.java
@@ -1,9 +1,0 @@
-package org.zalando.nakadi.exceptions.runtime;
-
-public class NoStorageException extends MyNakadiRuntimeException1 {
-
-    public NoStorageException(final String msg) {
-        super(msg);
-    }
-
-}