+++ src/main/java/org/zalando/nakadi/exceptions/runtime/RebalanceConflictException.java
@@ -1,6 +1,6 @@
 package org.zalando.nakadi.exceptions.runtime;
 
-public class RebalanceConflictException extends MyNakadiRuntimeException1 {
+public class RebalanceConflictException extends NakadiBaseException {
 
     public RebalanceConflictException(final String msg) {
         super(msg);