+++ src/main/java/org/zalando/nakadi/exceptions/runtime/TopicCreationException.java
@@ -1,6 +1,6 @@
 package org.zalando.nakadi.exceptions.runtime;
 
-public class TopicCreationException extends MyNakadiRuntimeException1 {
+public class TopicCreationException extends NakadiBaseException {
 
     public TopicCreationException(final String msg) {
         super(msg);