+++ src/main/java/org/zalando/nakadi/exceptions/runtime/WrongInitialCursorsException.java
@@ -1,6 +1,6 @@
 package org.zalando.nakadi.exceptions.runtime;
 
-public class WrongInitialCursorsException extends MyNakadiRuntimeException1 {
+public class WrongInitialCursorsException extends NakadiBaseException {
 
     public WrongInitialCursorsException(final String msg) {
         super(msg);