+++ src/main/java/org/zalando/nakadi/service/AuthorizationValidator.java
@@ -7,7 +7,7 @@ import org.zalando.nakadi.domain.EventType;
 import org.zalando.nakadi.domain.EventTypeResource;
 import org.zalando.nakadi.domain.Subscription;
 import org.zalando.nakadi.domain.ValidatableAuthorization;
-import org.zalando.nakadi.exceptions.InternalNakadiException;
+import org.zalando.nakadi.exceptions.runtime.InternalNakadiException;
 import org.zalando.nakadi.exceptions.runtime.AccessDeniedException;
 import org.zalando.nakadi.exceptions.runtime.ServiceTemporarilyUnavailableException;
 import org.zalando.nakadi.exceptions.runtime.UnableProcessException;