+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Fixed
+- Upgraded dependencies
+- Refactored exceptions
+
 ## [2.8.3] - 2018-08-01
 
 ### Fixed